+++ tests/parse_flags.c
@@ -32,7 +32,7 @@ int parse_flags(int argc, char **argv)
 	for (arg_idx = 1; arg_idx < argc ; arg_idx++)
 	{
 		int jj;
-		for (jj = 0; jj < NELEM(format_args); jj++)
+		for (jj = 0; jj < (int)NELEM(format_args); jj++)
 		{
 			if (strcasecmp(argv[arg_idx], format_args[jj].arg) == 0)
 			{
